+++ src/Core/Baskets/Services/BasketService.php
@@ -338,9 +338,11 @@ class BasketService extends BaseService
             $user = $this->getByHashedId($user);
         }
 
-        $basket = $user->latestBasket->load($includes);
+        if ($user->latestBasket) {
+            $basket = $user->latestBasket->load($includes);
+        }
 
-        if ($basket) {
+        if (!empty($basket)) {
             if ($basket->order && ! $basket->order->placed_at || ! $basket->order) {
                 return $this->factory->init($basket)->get();
             }