+++ CHANGELOG.md
@@ -1,3 +1,64 @@
+#0.2.13
+
+- [improved] Improved handling of order reference increment
+
+#0.2.12
+
+- [fixed] Fixed issue with installer due to country column change
+- [added] Added config for order table columns in the hub
+- [changed] Order settings now pass through all config
+
+#0.2.11
+
+- [changed] Changed plugin loader to not require file extension, this was causing some issues on specific nginx servers.
+- [fixed] Allow tracking no to be set to null on orders
+
+#0.2.10
+
+- [fixed] Fixed issue where updating shipping price zeroed out order totals.
+- [changed] Updated the set shipping price method for DI.
+- [fixed] Fixed offline payment driver, was a bit outdated.
+
+#0.2.7
+
+- [fix] Remove left over calls to countries json columns
+
+#0.2.6
+
+- [added] You can now specify search config for the client.
+
+``` php
+    // Rest of getcandy.php config
+    'search' => [
+        'client_config' => [
+            'elastic' => [
+                'host' => null,
+                'port' => null,
+                'path' => null,
+                'url' => null,
+                'proxy' => null,
+                'transport' => null,
+                'persistent' => true,
+                'timeout' => null,
+                'connections' => [], // host, port, path, timeout, transport, compression, persistent, timeout, username, password, config -> (curl, headers, url)
+                'roundRobin' => false,
+                'log' => false,
+                'retryOnConflict' => 0,
+                'bigintConversion' => false,
+                'username' => null,
+                'password' => null,
+            ]
+        ],
+    ]
+```
+
+#v0.2.5
+
+- [fixed] Fixed order notes and customer reference not saving on order.
+- [fixed] Fixed exception when retrieving the current basket if the authenticated user never had one before.
+- [changed] Don't round tiered pricing, comes back as three decimal places.
+- [changed] The country name column in the database is now a text field, we aim to support translation files instead for this.
+
 #v0.2.4
 
 - [fixed] Fixed issue with order searching caused by changing variant to description on order lines.